version 3

+++ src/main/java/com/banking/banking/Repo/AccountRepository.java
@@ -1,8 +1,7 @@
 package com.banking.banking.Repo;
 
-
 import com.banking.banking.Model.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
 }
